fix to getLargestProbEl -- long long ints required when more than 29 qubits are stored per node

+++ 3sat/archer/35qubits_32ranksx24threads/3SATSolver.c
@@ -128,7 +128,7 @@ void createOrClearFile(char *filename) {
 
 
 void extractSolution(MultiQubit qubits, int *solBits, REAL *solProb) {
-	int solInt = 0;
+	long long int solInt = 0;
 	getLargestProbEl(qubits, solProb, &solInt);
 	convertToBinary(solInt, qubits.numQubits, solBits);
 }
